+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes of GetSimple Legacy CMS will be documented in this file.
 ### Added
 
 - Display log file as text on the page `/admin/log.php`. Add GET parameter `view=text` to URL.
+- Display content of the PHP errors log file. Use URL `/admin/log.php?log=errorlog.txt&view=text`.
 
 ## [2025.1.1] - 2025.01.28